requirements.txt
lxml
tmdbv3api
python-dotenv
nltk
gunicorn
requests
Flask-Migrate
beautifulsoup4
Flask
flask-sqlalchemy
psycopg2
pandas
numpy
scikit-learn
urllib3
faiss-cpu
lief